+++ examples/ex7.cpp
@@ -218,7 +218,7 @@ class physics_test
     {
         _win.set_title(title);
     }
-    void solve(const double frame_time, const double damping)
+    void solve(const float frame_time, const float damping)
     {
         // Solve the simulation
         _simulation.solve(frame_time, damping);
@@ -271,7 +271,7 @@ int test_render_loop()
     min::loop_sync sync(frames);
 
     // User can close with Q or use window manager
-    double frame_time = 0.0;
+    float frame_time = 0.0;
     while (!test.is_closed())
     {
         for (int i = 0; i < frames; i++)
@@ -282,7 +282,7 @@ int test_render_loop()
             // Update rigid bodies in simulation
             for (int i = 0; i < 30; i++)
             {
-                test.solve(frame_time / 30.0, 0.01);
+                test.solve(frame_time / 30.0f, 0.01f);
             }
 
             // Update the bodies buffer information
@@ -298,7 +298,7 @@ int test_render_loop()
             test.window_update();
 
             // Calculate needed delay to hit target
-            frame_time = sync.sync();
+            frame_time = static_cast<float>(sync.sync());
         }
 
         // Calculate the number of 'average' frames per second
